store/tikv: move tikv driver out
